Y. Tablas de metas listo

+++ database/migrations/2022_03_18_022942_create_objetivo_educacionals_table.php
@@ -17,7 +17,7 @@ return new class extends Migration
             $table->id();
             $table->string('descripcion');
             $table->bigInteger('idCarrera')->unsigned();
-            $table->bigInteger('meta');
+            $table->float('meta', 8, 2);
             $table->timestamps();
             $table->foreign('idCarrera')->references('id')->on('carreras');//->onDelete('cascade');
         });
